+++ Utils/Runtime/ProjectileMath.cs
@@ -13,15 +13,15 @@ namespace CheesePie.Utils {
 		/// and an intersection point. Will always choose the velocity with the smaller y value.
 		/// </summary>
 		/// <param name="point">Delta from initial point</param>
-		/// <param name="gravity">Vertical gravity (Downwards is negative)</param>
+		/// <param name="yForce">Vertical force (Downwards is negative)</param>
 		/// <returns>The Initial velocity</returns>
-		public static Vector2? FindInitialVelocity(float initialSpeed, Vector2 point, float gravity) {
+		public static Tuple<Vector2, Vector2> FindInitialVelocity(float initialSpeed, Vector2 point, float yForce) {
 			// Using these equations:
 			//   initialVelocity.x = initialSpeed * cos(theta) * time
-			//   initialVelocity.y = 0.5 * gravity * time^2 + initialSpeed * sin(theta) * time
+			//   initialVelocity.y = 0.5 * yForce * time^2 + initialSpeed * sin(theta) * time
 			// And using the trig identity 1/cos^2(theta) = tan^2(theta) + 1
 			// We can calculate the a, b and c values of a quadratic equation to solve for tan(theta)
-			float a = 0.5f * gravity * point.x * point.x * (1f / (initialSpeed * initialSpeed));
+			float a = 0.5f * yForce * point.x * point.x * (1f / (initialSpeed * initialSpeed));
 			float b = point.x;
 			float c = a - point.y;
 			Tuple<float, float> roots = CPMath.SolveQuadratic(a, b, c);
@@ -41,11 +41,7 @@ namespace CheesePie.Utils {
 			Vector2 velocity1 = new Vector2(Mathf.Cos(angle1), Mathf.Sin(angle1)) * initialSpeed;
 			Vector2 velocity2 = new Vector2(Mathf.Cos(angle2), Mathf.Sin(angle2)) * initialSpeed;
 
-			if (velocity1.y < velocity2.y) {
-				return velocity1;
-			} else {
-				return velocity2;
-			}
+			return new Tuple<Vector2, Vector2>(velocity1, velocity2);
 		}
 	}
 }