+++ Utils/CHANGELOG.md
@@ -8,12 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CPMath (general math functions)
 - ProjectileMath (math functions related to projectiles)
-- Tests for FiniteStateMachine
+- Added swizzle extension functions for Vector3 to Vector2
+- Editor tests for FiniteStateMachine, CPMath, Extensions, ProjectileMath
+- Play mode tests for MonoBehaviour extension methods
 
 ### Changed
 - Better tests for ObjectPool
 
 ### Removed
+- Removed area and perimeter extension functions
+- Removed Divide extension functions for Vector2Int
 
 ## [0.0.1] - 2021-01-24
 ### Added